move the hook aside /markdown page component
